+++ common/src/main/java/dev/detpikachu/buildersvoid/ModConstants.java
@@ -9,6 +9,9 @@ public class ModConstants {
 	public static final String MOD_ID = "buildersvoid";
 	public static final Logger LOG = LoggerFactory.getLogger(MOD_ID);
 
+    public static final int VOID_HEIGHT = 50;
+    public static final int VOID_SPACING = 10000;
+
     public static ResourceLocation id(String path) {
 
         return new ResourceLocation(MOD_ID, path);